linear-classify.md - regularizer: SQUARED l2 norm
isnt the regularizer we are using for W the SQUARED l2 norm? because if we were to use the regular l2 norm we would need to take the root over the entire sum of squares (i.e. R(W) = ||W||^2). but we are not doing that. it is just a small detail, i know, but it has been bugging me. best(: peglegpete